requirements.txt
Flask==3.0.0
requests==2.31.0
dnspython==2.4.2
pyOpenSSL==23.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.1.0
cryptography==41.0.7